First iteration of streamlit dashboard (CRUD entry)

+++ sql/stage/sessions/000_create_tables.sql
@@ -21,7 +21,7 @@ CREATE TABLE session_batch_inventory
 	product_name    TEXT,
 	vendor_name     TEXT,
 	production_date DATE,
-	quantity_used   INT,
+	quantity_used   NUMERIC,
 	quantity_output NUMERIC,
 	role            TEXT,
 	batch_code      TEXT,
